refactor: update protocol_rpc with refactored consensus_algorithm
